requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0

# Vector database
chromadb>=0.4.15
sentence-transformers>=2.2.2

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Redis for feedback queue
redis>=5.0.0
celery>=5.3.0

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data processing
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Production dependencies
gunicorn>=21.2.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment management
python-dotenv>=1.0.0
